Logged in vs Logged out mode v1
